feat(build): add revision arg and usage

+++ build.sh
@@ -3,18 +3,30 @@
 # architecture is checked to be one of the official architectures in order to
 # guard against typosquatting and name squatting attacks on the Docker Hub.
 
-arch="$*"
-case "$arch" in
-    amd64|arm32v6|arm64v8|i386|ppc64le|s390x)
-        ;;
-    '')
-        echo 'no target architecture' > /dev/stderr
-        exit 1
-        ;;
-    *)
-        echo "unknown architecture: \"$arch\"" > /dev/stderr
-        exit 1
-esac
+archs="(amd64|arm32v6|arm64v8|i386|ppc64le|s390x)"
+
+usage() {
+    echo "usage: $(basename "$0") $archs [revision]" > /dev/stderr
+}
+
+
+arch="$1"
+revision="${2:-HEAD}"
+if ! [ -z "$3" ]; then
+    usage
+    exit 1
+fi
+
+
+if [ -z "$arch" ]; then
+    echo 'no target architecture' > /dev/stderr
+    usage
+    exit 1
+elif ! echo "$arch" | grep -E -q "$archs"; then
+    echo "unknown architecture: \"$arch\"" > /dev/stderr
+    usage
+    exit 1
+fi
 echo "target architecture is $arch"
 
 
@@ -35,10 +47,12 @@ docker run \
      && cd \
      && git clone https://github.com/msteveb/jimtcl \
      && cd jimtcl/ \
+     && git checkout "$1" \
      && ./configure --full --with-ext=sqlite3 \
      && env LDLIBS=-static make \
      && cp jimsh \
           "/inbox/jimsh-$(./jimsh --version)-$(git rev-parse HEAD \
                                                | cut -c 1-10)-$0"
     ' \
-    "$arch"
+    "$arch" \
+    "$revision"
